requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
alembic==1.9.2
sqlmodel==0.0.8
asyncpg==0.27.0
python-dotenv==0.21.1
pydantic==1.10.4
httpx==0.23.3
pytest==7.2.1
pytest-asyncio==0.20.3
aiosqlite==0.18.0
aioredis==2.0.1
pre-commit==3.0.2